+++ src/gui/NodeModule.cpp
@@ -467,9 +467,9 @@ NodeModule::property_changed(const URI& key, const Atom& value)
 	const URIs& uris = app().uris();
 	if (value.type() == uris.forge.Float) {
 		if (key == uris.ingen_canvasX) {
-			move_to(value.get<float>(), get_y());
+			move_to(static_cast<double>(value.get<float>()), get_y());
 		} else if (key == uris.ingen_canvasY) {
-			move_to(get_x(), value.get<float>());
+			move_to(get_x(), static_cast<double>(value.get<float>()));
 		}
 	} else if (value.type() == uris.forge.Bool) {
 		if (key == uris.ingen_polyphonic) {
